+++ src/config/constants.ts
@@ -4,5 +4,3 @@
 
 export const SERVER_NAME = 'firefox-devtools';
 export const SERVER_VERSION = '0.2.5';
-
-export const DEFAULT_HEADLESS = false;